RC4

+++ PC/src/Entities/Coin.cpp
@@ -59,8 +59,8 @@ bool Coin::UpdateWhenNoCoins()
 void Coin::Draw()
 {
 	_g->BlitFrame(this->_animPlayer.GetCurrentFrame(), 
-		          (this->_x * this->_tileSize.x), 
-		          (this->_y * this->_tileSize.y), 
+		          (int)(this->_x * this->_tileSize.x), 
+		(int)(this->_y * this->_tileSize.y),
 				  this->_tileSize.x,
 				  this->_tileSize.y,
 				  false, false);
